+++ changelog.md
@@ -1,5 +1,13 @@
+## VERSION 2.Alpha
 	### UI
 		+	improved responsiveness
+		+	easier to understand statistics
+		+	added description to history
+		+	property sell separated for ease of use
+		+	modified navbar and footer colors
+	### IN CONSTRUCTION
+		+	user specific transactions
+
 ## VERSION 1.4.0.0
 	### UI
 		+	added index column to history